PreviewImage component introduced; switching between different types of preview bug is fixed;
