+++ packages/template-block/src/components/TemplatePreview.tsx
@@ -1,24 +1,15 @@
 /* (c) Copyright Frontify Ltd., all rights reserved. */
 
-import { AppBridgeBlock, Template, useBlockAssets, useEditorState } from '@frontify/app-bridge';
-import { IconArrowSync, IconPlus24, IconSpeechBubbleQuote20, merge } from '@frontify/fondue';
-import { useState } from 'react';
-import {
-    PreviewType,
-    Settings,
-    previewDisplayValues,
-    previewHeightValues,
-    previewImageAnchoringValues,
-    textPositioningToFlexDirection,
-} from '../types';
+import { AppBridgeBlock, Template, useBlockAssets } from '@frontify/app-bridge';
+import { IconPlus24 } from '@frontify/fondue';
+import { Settings, previewHeightValues, textPositioningToFlexDirection } from '../types';
 import {
     BlockInjectButton,
-    BlockItemWrapper,
     getBackgroundColorStyles,
     radiusStyleMap,
     toRgbaString,
 } from '@frontify/guideline-blocks-settings';
-import { EditAltTextFlyout } from '@frontify/guideline-blocks-shared';
+import { PreviewImage } from './PreviewImage';
 
 export type TemplatePreviewProps = {
     appBridge: AppBridgeBlock;
@@ -36,10 +27,6 @@ export const TemplatePreview = ({
     onOpenTemplateChooser,
 }: TemplatePreviewProps) => {
     const { blockAssets } = useBlockAssets(appBridge);
-    const isEditing = useEditorState(appBridge);
-    const [showAltTextMenu, setShowAltTextMenu] = useState(false);
-    const [localAltText, setLocalAltText] = useState<string | undefined>(blockSettings.altText);
-    const [isHovered, setIsHovered] = useState(false);
 
     const {
         hasBackgroundTemplatePreview,
@@ -52,12 +39,9 @@ export const TemplatePreview = ({
         radiusValue_templatePreview,
         radiusChoice_templatePreview,
         textRatio,
-        preview,
         isPreviewHeightCustom,
         previewHeightSimple,
         previewHeightCustom,
-        previewDisplay,
-        previewImageAnchoring,
         textPositioning,
     } = blockSettings;
 
@@ -93,58 +77,12 @@ export const TemplatePreview = ({
                         height,
                     }}
                 >
-                    <img
-                        data-test-id="template-block-preview-img"
-                        src={
-                            preview === PreviewType.Custom && previewCustom
-                                ? previewCustom[0].previewUrl
-                                : template?.previewUrl
-                        }
-                        className={merge(['tw-relative tw-w-full tw-h-full', previewDisplayValues[previewDisplay]])}
-                        style={{
-                            objectPosition: previewImageAnchoring
-                                ? previewImageAnchoringValues[previewImageAnchoring]
-                                : 'center',
-                        }}
-                        width={preview === PreviewType.Custom && previewCustom ? previewCustom[0].width : 'auto'}
-                        height={preview === PreviewType.Custom && previewCustom ? previewCustom[0].height : 'auto'}
-                        alt={blockSettings.altText || undefined}
-                        onMouseEnter={() => setIsHovered(true)}
-                        onMouseLeave={() => setIsHovered(false)}
-                        onBlur={() => setIsHovered(false)}
+                    <PreviewImage
+                        appBridge={appBridge}
+                        blockSettings={blockSettings}
+                        updateBlockSettings={updateBlockSettings}
+                        onOpenTemplateChooser={onOpenTemplateChooser}
                     />
-                    {isEditing && (
-                        <div className="tw-absolute tw-top-5 tw-right-0 tw-flex tw-justify-end tw-pt-3 tw-mx-3">
-                            <BlockItemWrapper
-                                shouldHideWrapper={!isEditing}
-                                shouldBeShown={isHovered || showAltTextMenu}
-                                toolbarItems={[]}
-                                toolbarFlyoutItems={[
-                                    [
-                                        {
-                                            title: 'Set alt text',
-                                            onClick: () => setShowAltTextMenu(true),
-                                            icon: <IconSpeechBubbleQuote20 />,
-                                        },
-                                        {
-                                            title: 'Replace template',
-                                            onClick: () => onOpenTemplateChooser(),
-                                            icon: <IconArrowSync />,
-                                        },
-                                    ],
-                                ]}
-                            >
-                                <EditAltTextFlyout
-                                    setShowAltTextMenu={setShowAltTextMenu}
-                                    showAltTextMenu={showAltTextMenu}
-                                    setLocalAltText={setLocalAltText}
-                                    defaultAltText={blockSettings.altText}
-                                    onSave={() => updateBlockSettings({ altText: localAltText || undefined })}
-                                    localAltText={localAltText}
-                                />
-                            </BlockItemWrapper>
-                        </div>
-                    )}
                 </div>
             ) : (
                 <div style={{ height }}>
